# vulkan: im2col and matmul optimizations for stable diffusion