--- a/ggml/src/ggml-vulkan/vulkan-shaders/im2col.comp
+++ b/ggml/src/ggml-vulkan/vulkan-shaders/im2col.comp
@@ -26,7 +26,7 @@ layout (push_constant) uniform parameter
 
 layout(constant_id = 0) const uint BLOCK_SIZE = 32;
 
-const uint NUM_ITER = 256 / BLOCK_SIZE;
+const uint NUM_ITER = 512 / BLOCK_SIZE;
 
 layout(local_size_x_id = 0, local_size_y = 1, local_size_z = 1) in;